+++ code/index/java/nu/marginalia/index/model/SearchContext.java
@@ -18,8 +18,8 @@ import nu.marginalia.api.searchquery.model.results.PrototypeRankingParameters;
 import nu.marginalia.index.CombinedIndexReader;
 import nu.marginalia.index.reverse.IndexLanguageContext;
 import nu.marginalia.index.reverse.query.IndexSearchBudget;
-import nu.marginalia.ranking.data.SearchSet;
-import nu.marginalia.ranking.data.ConnectivityView;
+import nu.marginalia.ranking.set.SearchSet;
+import nu.marginalia.ranking.connectivity.ConnectivityView;
 import nu.marginalia.language.keywords.KeywordHasher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
