+++ code/index/java/nu/marginalia/index/searchset/ConnectivitySets.java
@@ -5,8 +5,8 @@ import com.google.inject.Singleton;
 import it.unimi.dsi.fastutil.ints.Int2ByteOpenHashMap;
 import nu.marginalia.db.DomainRankingSetsService;
 import nu.marginalia.index.IndexFactory;
-import nu.marginalia.ranking.data.ConnectivityMapFile;
-import nu.marginalia.ranking.data.ConnectivityView;
+import nu.marginalia.ranking.connectivity.ConnectivityMapFile;
+import nu.marginalia.ranking.connectivity.ConnectivityView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
