+++ code/index/test/nu/marginalia/index/CombinedIndexReaderTest.java
@@ -14,7 +14,7 @@ import nu.marginalia.index.reverse.construction.DocIdRewriter;
 import nu.marginalia.index.reverse.construction.full.FullIndexConstructor;
 import nu.marginalia.index.reverse.construction.prio.PrioIndexConstructor;
 import nu.marginalia.index.reverse.query.IndexSearchBudget;
-import nu.marginalia.ranking.data.DomainRankings;
+import nu.marginalia.ranking.DomainRankings;
 import nu.marginalia.language.keywords.KeywordHasher;
 import nu.marginalia.linkdb.docs.DocumentDbReader;
 import nu.marginalia.linkdb.docs.DocumentDbWriter;
