+++ code/processes/ranking-process/java/nu/marginalia/ranking/ConnectivitySetsCalculator.java
@@ -12,10 +12,10 @@ import it.unimi.dsi.fastutil.ints.IntSet;
 import nu.marginalia.IndexLocations;
 import nu.marginalia.api.linkgraph.AggregateLinkGraphClient;
 import nu.marginalia.db.DomainRankingSetsService;
-import nu.marginalia.ranking.data.RankingSearchSet;
-import nu.marginalia.ranking.data.ConnectivityMapFile;
-import nu.marginalia.ranking.data.ConnectivityView;
-import nu.marginalia.ranking.data.DomainSetConnectivity;
+import nu.marginalia.ranking.set.RankingSearchSet;
+import nu.marginalia.ranking.connectivity.ConnectivityMapFile;
+import nu.marginalia.ranking.connectivity.ConnectivityView;
+import nu.marginalia.ranking.connectivity.DomainSetConnectivity;
 import nu.marginalia.storage.FileStorageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
