+++ code/processes/ranking-process/java/nu/marginalia/ranking/PrimaryRankingsCalculator.java
@@ -5,12 +5,11 @@ import it.unimi.dsi.fastutil.ints.Int2IntOpenHashMap;
 import nu.marginalia.IndexLocations;
 import nu.marginalia.db.DomainRankingSetsService;
 import nu.marginalia.db.DomainTypes;
-import nu.marginalia.domainranking.PageRankDomainRanker;
-import nu.marginalia.domainranking.accumulator.RankingResultHashMapAccumulator;
+import nu.marginalia.ranking.domains.PageRankDomainRanker;
+import nu.marginalia.ranking.domains.accumulator.RankingResultHashMapAccumulator;
 import nu.marginalia.process.ProcessConfiguration;
 import nu.marginalia.process.control.ProcessEventLog;
-import nu.marginalia.ranking.data.ConnectivityView;
-import nu.marginalia.ranking.data.DomainRankings;
+import nu.marginalia.ranking.connectivity.ConnectivityView;
 import nu.marginalia.storage.FileStorageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
