+++ code/processes/ranking-process/java/nu/marginalia/ranking/SecondaryRankingsCalculator.java
@@ -6,10 +6,10 @@ import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import nu.marginalia.IndexLocations;
 import nu.marginalia.db.DomainRankingSetsService;
 import nu.marginalia.db.DomainTypes;
-import nu.marginalia.domainranking.PageRankDomainRanker;
-import nu.marginalia.domainranking.accumulator.RankingResultHashSetAccumulator;
+import nu.marginalia.ranking.domains.PageRankDomainRanker;
+import nu.marginalia.ranking.domains.accumulator.RankingResultHashSetAccumulator;
 import nu.marginalia.process.control.ProcessEventLog;
-import nu.marginalia.ranking.data.RankingSearchSet;
+import nu.marginalia.ranking.set.RankingSearchSet;
 import nu.marginalia.storage.FileStorageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
