+++ code/processes/ranking-process/java/nu/marginalia/ranking/domains/accumulator/RankingResultAccumulator.java
@@ -1,4 +1,4 @@
-package nu.marginalia.domainranking.accumulator;
+package nu.marginalia.ranking.domains.accumulator;
 
 public interface RankingResultAccumulator<T> {
     void add(int domainId, int rank);
