+++ code/processes/ranking-process/model/java/nu/marginalia/ranking/set/SearchSet.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.data;
+package nu.marginalia.ranking.set;
 
 import it.unimi.dsi.fastutil.ints.IntList;
