+++ code/processes/ranking-process/model/test/nu/marginalia/ranking/set/RankingSearchSetTest.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.data;
+package nu.marginalia.ranking.set;
 
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import org.junit.jupiter.api.Test;
