+++ code/processes/ranking-process/test/nu/marginalia/ranking/domains/RankingAlgorithmsContainerTest.java
@@ -1,4 +1,4 @@
-package nu.marginalia.domainranking;
+package nu.marginalia.ranking.domains;
 
 
 import com.zaxxer.hikari.HikariConfig;
