+++ code/processes/ranking-process/test/nu/marginalia/ranking/domains/TestGraphSourceForLinkData.java
@@ -1,4 +1,4 @@
-package nu.marginalia.domainranking;
+package nu.marginalia.ranking.domains;
 
 import nu.marginalia.array.LongArrayFactory;
 import nu.marginalia.domaingraph.DomainGraph;
@@ -6,6 +6,7 @@ import nu.marginalia.domaingraph.DomainGraphBuilder;
 import nu.marginalia.domaingraph.GraphSource;
 import org.apache.commons.lang3.StringUtils;
 
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -13,9 +14,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class TestGraphSourceForInvertedLinkData implements GraphSource {
+public class TestGraphSourceForLinkData implements GraphSource {
     // The data is available at
     // https://downloads.marginalia.nu/link-test-data.tar.gz
+
     private static Path domainDataPath = Paths.get("/home/vlofgren/Exports/Links/domains.export.tsv");
     private static Path[] linksDataPaths = new Path[] {
             Paths.get("/home/vlofgren/Exports/Links/domain-links-1.dat"),
@@ -35,10 +37,18 @@ public class TestGraphSourceForInvertedLinkData implements GraphSource {
         return Files.exists(domainDataPath) && Files.exists(linksDataPaths[0]);
     }
 
+    private boolean loaded = false;
+
     private Map<Integer, String> idToName = new HashMap<>();
 
     public String getName(int id) {
-        return idToName.get(id);
+        if (!loaded) throw new IllegalStateException("Graph not loaded, run getGraph() first!");
+
+        String name = idToName.get(id);
+        if (null == name) {
+            return "id:"+Integer.toString(id);
+        }
+        return name;
     }
 
     @Override
@@ -46,16 +56,25 @@ public class TestGraphSourceForInvertedLinkData implements GraphSource {
         DomainGraphBuilder builder = DomainGraphBuilder.directed();
         idToName = new HashMap<>();
 
+        try {
+            Files.readAllLines(domainDataPath).forEach(line -> {
+                String[] parts = line.split("\t");
+                if (!Character.isDigit(parts[0].charAt(0)))
+                    return;
+                idToName.put(Integer.parseInt(parts[0]), parts[1]);
+            });
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
         try (var stream = Files.lines(domainDataPath)) {
             stream.skip(1)
                     .mapMultiToInt((line, c) -> {
                         String[] parts = StringUtils.split(line, '\t');
                         int id = Integer.parseInt(parts[0]);
-                        String name = parts[1];
                         int node_affinity = Integer.parseInt(parts[3]);
                         if (node_affinity > 0) {
                             c.accept(id);
-                            idToName.put(id, name);
                         }
                     })
                     .forEach(builder::addVertex);
@@ -64,7 +83,7 @@ public class TestGraphSourceForInvertedLinkData implements GraphSource {
             throw new RuntimeException(e);
         }
 
-        return builder.build(consumer -> {
+        DomainGraph graph = builder.build(consumer -> {
             for (var path : linksDataPaths) {
                 try (var data = LongArrayFactory.mmapForReadingConfined(path)) {
                     data.forEach(0, data.size(), (pos, val) -> {
@@ -72,8 +91,7 @@ public class TestGraphSourceForInvertedLinkData implements GraphSource {
 
                         int src = (int) (val >>> 32);
                         int dest = (int) (val & 0xFFFF_FFFFL);
-                        // Invert the edge
-                        consumer.accept(dest, src);
+                        consumer.accept(src, dest);
                     });
                 }
                 catch (Exception e) {
@@ -81,5 +99,8 @@ public class TestGraphSourceForInvertedLinkData implements GraphSource {
                 }
             }
         });
+
+        loaded = true;
+        return graph;
     }
 }
